requirements.txt
fedlab==1.3.0
hydra-core==1.3.2
matplotlib==3.7.5
networkx==3.1
numpy==1.24.3
omegaconf==2.3.0
pandas==2.0.3
scikit_learn==1.3.2
torch==1.13.1
pyarrow==17.0.0